Refactor GameOptions and main menu events to use better events and reduce coupling. Make GameOptions not a component anymore
