+++ source/core/src/main/com/csse3200/game/GdxGame.java
@@ -3,7 +3,7 @@ package com.csse3200.game;
 import com.badlogic.gdx.Game;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Screen;
-import com.csse3200.game.components.mainmenu.GameOptions;
+import com.csse3200.game.options.GameOptions;
 import com.csse3200.game.files.UserSettings;
 import com.csse3200.game.screens.MainGameScreen;
 import com.csse3200.game.screens.MainMenuScreen;
