+++ source/core/src/main/com/csse3200/game/screens/MainMenuScreen.java
@@ -5,7 +5,6 @@ import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.csse3200.game.GdxGame;
 import com.csse3200.game.components.mainmenu.MainMenuActions;
 import com.csse3200.game.components.mainmenu.MainMenuDisplay;
-import com.csse3200.game.components.mainmenu.GameOptions;
 import com.csse3200.game.entities.Entity;
 import com.csse3200.game.entities.EntityService;
 import com.csse3200.game.entities.factories.RenderFactory;
@@ -99,8 +98,7 @@ public class MainMenuScreen extends ScreenAdapter {
         Entity ui = new Entity();
         ui.addComponent(new MainMenuDisplay())
                 .addComponent(new InputDecorator(stage, 10))
-                .addComponent(new MainMenuActions(game))
-                .addComponent(new GameOptions());
+                .addComponent(new MainMenuActions(game));
         ServiceLocator.getEntityService().register(ui);
     }
 }
